Add binary, linked list bug fix, mem enhancement
Found a bug on linked lists that saves LIDAR data for elevation discovery, which will crash MATLAB. Fixed this, and enhanced memory management on serial use so that after consecutive usage of this program, MATLAB will not occupy too much memory for saving data (but unused instead).
The missing LIDAR file will not be shown in MATLAB's warning page, written to a warning.log file instead. consecutive misses of a file will only show once.
For consecutive usage a sample MATLAB code is shown.